+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.0
 * RefLinkRequest.id (int) -> linksequenceId (long)
 * RoadSysRefRequest.roadCategory,.phase changed to enums
+* SegmentedLink.linkType, .detailLevel, .roadType changed to enums
 
 ## 1.3.0
 * Fixed bug when parsing RoadSysRef of road segments in object.